Update SentEmailsServiceProvider and migration files
This commit updates the SentEmailsServiceProvider by adding a method to load migrations from the specified directory. It renames and modifies the 'create_sent_emails_table' migration file using the date as part of the filename and adds a new migration for creating 'sent_emails_attachments' table. Anonymous classes are used for migration classes to simplify the code.